Starting users modele, minor fixes of replays and replays comments

+++ app/Http/Livewire/Replay/Delete.php
@@ -4,20 +4,28 @@ namespace App\Http\Livewire\Replay;
 
 use Livewire\Component;
 use App\Models\Replays\Replay;
+use App\Policies\ReplayPolicy;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class Delete extends Component
 {
+    use AuthorizesRequests;
+
     public $modalVisibility = false;
     public $replay;
 
     public function loadModal(Replay $replay)
     {
+        $this->authorize(ReplayPolicy::DELETE, $this->replay);
+
         $this->replay = $replay;
         $this->modalVisibility = true;
     }
 
     public function deleteReplay()
     {
+        $this->authorize(ReplayPolicy::DELETE, $this->replay);
+
         $this->replay->delete();
         $this->modalVisibility = false;
         return redirect()->route('replays.index')->with('success', 'Replay delete.');
